shuffle letters to fix issue where there is never a left branch (Left=null)
With few trees, we never even split left, because the letter variables happened
to be in order roughly by frequency instead of randomly distributed in the indexedVariables
array and variables map. This is because as you go through allChars during training,
the likelihood that a character was already seen gets lower and lower the further
you go. So you end up with rare letters distributed far to the end of the indexedVariables
array. Now they get shuffled, and there are typically always trees that have Left
branches.